ADC/Tag2AdDemoSite-FrontEndDev-v0.8-2/requirements.txt
flask>=0.12.1
tornado>=4.5.1
msgpack-python>=0.4.8